create plan, regitration and checkin

+++ src/app/models/Student.js
@@ -7,7 +7,7 @@ class Student extends Model {
         name: Sequelize.STRING,
         email: Sequelize.STRING,
         age: Sequelize.INTEGER,
-        weight: Sequelize.FLOAT,
+        weight: Sequelize.INTEGER,
         height: Sequelize.INTEGER,
       },
       {
